+++ R/swfscDAS-package.R
@@ -12,7 +12,7 @@
 #' @author Sam Woodman \email{sam.woodman@@noaa.gov}
 #' @seealso \url{https://swfsc.github.io/swfscDAS/}
 #'
-#' @importFrom dplyr add_row arrange between bind_cols bind_rows case_when desc distinct everything filter full_join group_by if_else lag left_join mutate n n_distinct reframe right_join select slice starts_with summarise ungroup
+#' @importFrom dplyr across add_row arrange between bind_cols bind_rows case_when desc distinct everything filter full_join group_by if_else lag left_join mutate n n_distinct pull reframe right_join select slice starts_with summarise ungroup
 #' @importFrom lubridate year month day tz
 #' @importFrom magrittr %>%
 #' @importFrom methods setOldClass