+++ NEWS.md
@@ -1,18 +1,26 @@
 # swfscDAS 0.7.0.9000
 
-* Fix bug in `das_process` of incorrect OffsetGMT values, stemming from `as.Date` using 'UTC' tz by default (#16)
+* Fixed bug in `das_process` of incorrect OffsetGMT values, stemming from `as.Date` using 'UTC' tz by default (#16)
+
+* Fixed bug in `das_segdata` to actually only use the last condition value for a segment of length zero (#27)
 
 * Added an exported function `das_spcodes_read` for consistent reading of shipboard SpCode files
 
 * Changed the argument name in `das_check` of the user-provided SpCodes file from `sp.codes` to `sp.codes.file`. Also added a `sp.codes.skip` argument to pass to `das_spcodes_read`
 
+* Added 'SwellDir' (Event V, Data 3) and 'WindDir' (Event W, Data 4) to the output of `das_process` (#21)
+
+* Changed behavior of `das_effort` when the `seg0.drop` argument is `TRUE`. Now, for segments with distance traveled <=0.1 km, either a) they are dropped if they have no associated sighting events are dropped, or b) if they have associated sightings they are kept and given a distance of 0.1km. This behavior matches historical SWFSC segment chopping logic (#22)
+
+* Added optional argument `gs.sp.min1` to `das_sight`, to allow the user to choose to set the minimum value for all non-NA species group size values to 1. This functionality is important for  group size calibration efforts. Users can also specify this argument in `das_effort`, where it is passed to `das_sight` when creating the sightinfo data frame (#26)
+
 * Moved example data files to an 'extdata' folder within 'inst'
 
-* Update link to NOAA Tech Memo describing DAS data; now (https://repository.library.noaa.gov/view/noaa/3174)[https://repository.library.noaa.gov/view/noaa/3174](#17)
+* Update link to NOAA Tech Memo describing DAS data; now [https://repository.library.noaa.gov/view/noaa/3174](#17)
 
-* Update roxygen to use markdown, fix documentation issues related to upgrade to roxygen2 v8.0
+* Update roxygen to use markdown, fix documentation issues related to upgrade to roxygen2 v8
 
-* Added the SpCodes.dat file from [`CruzPlot`](https://github.com/SWFSC/CruzPlot) to 'inst' to have available as a default
+* Added the SpCodes.dat file (from [`CruzPlot`](https://github.com/SWFSC/CruzPlot)) to 'inst' to have available as a default
 
 
 # swfscDAS 0.6.4